requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
qrcode==7.4.2
Pillow==10.0.0
Werkzeug==2.3.6
python-dotenv==1.0.0
